refactor(state): manage state in app
