+++ src/pages/about.tsx
@@ -1,18 +1,13 @@
-import { useEffect, useState } from 'react';
 import Link from 'next/link';
 
-const About = (): JSX.Element => {
-    const [nickname, setNickname] = useState('');
-    useEffect(() => {
-        const local = localStorage.getItem('nickname');
-        if (local) {
-            const stored = JSON.parse(local);
-            setNickname(stored.nickname);
-        }
-    });
+interface AboutProps {
+    nickname: string;
+}
+const About = (props: AboutProps): JSX.Element => {
+    const { nickname } = props;
     return (
         <div className="container">
-            <p>FOR FUCK SAKE</p>
+            <p>Cards Against Humanity Clone</p>
             <p>nickname: {nickname}</p>
             <Link href="/">
                 <a>Home</a>
